+++ packages/espressocash_app/lib/features/outgoing_link_payments/screens/share_link_screen.dart
@@ -1,12 +1,9 @@
-import 'package:barcode_widget/barcode_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:share/share.dart';
 
 import '../../../l10n/l10n.dart';
 import '../../../ui/app_bar.dart';
-import '../../../ui/button.dart';
-import '../../../ui/rounded_rectangle.dart';
+import '../../../ui/share_link.dart';
 import '../../../ui/theme.dart';
 import '../../authenticated/authenticated_navigator_key.dart';
 import '../../conversion_rates/widgets/extensions.dart';
@@ -42,44 +39,11 @@ class ShareLinkScreen extends StatelessWidget {
         appBar: CpAppBar(title: title),
         body: SafeArea(
           top: false,
-          child: Padding(
-            padding: const EdgeInsets.all(24),
-            child: CpRoundedRectangle(
-              backgroundColor: Colors.black,
-              scrollable: false,
-              padding: const EdgeInsets.symmetric(horizontal: 32, vertical: 40),
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  Text(
-                    context.l10n.scanToReceive(formattedAmount),
-                    maxLines: 3,
-                    textAlign: TextAlign.center,
-                    overflow: TextOverflow.ellipsis,
-                    style: const TextStyle(
-                      fontSize: 18,
-                      fontWeight: FontWeight.bold,
-                    ),
-                  ),
-                  Padding(
-                    padding: const EdgeInsets.symmetric(vertical: 40.0),
-                    child: BarcodeWidget(
-                      height: 250,
-                      barcode: Barcode.qrCode(),
-                      data: status.link.toString(),
-                      padding: EdgeInsets.zero,
-                      color: Colors.white,
-                    ),
-                  ),
-                  CpButton(
-                    text: context.l10n.share,
-                    width: double.infinity,
-                    size: CpButtonSize.big,
-                    onPressed: () => Share.share(message),
-                  ),
-                ],
-              ),
-            ),
+          child: ShareCodeWidget(
+            title: context.l10n.scanToReceive,
+            amount: formattedAmount,
+            qrCode: status.link.toString(),
+            shareText: message,
           ),
         ),
       ),
