+++ packages/espressocash_app/lib/l10n/intl_en.arb
@@ -368,36 +368,16 @@
   "@requestPaymentCreated": {},
   "requestPaymentReceived": "Money received",
   "@requestPaymentReceived": {},
-  "requestPaymentPending": "Waiting for payment",
+  "requestPaymentPending": "Payment status: Unpaid",
   "@requestPaymentPending": {},
   "requestDeletePaymentTitle": "Delete Request",
   "@requestDeletePaymentTitle": {},
-  "requestPaymentStatus": "Payment status: {status}",
-  "@requestPaymentStatus": {
-    "placeholders": {
-      "status": {
-        "type": "String"
-      }
-    }
-  },
   "save": "Save",
   "@save": {},
-  "scanToReceive": "Scan QR code to receive\n{amount}",
-  "@scanToReceive": {
-    "placeholders": {
-      "amount": {
-        "type": "String"
-      }
-    }
-  },
-  "scanToSend": "Scan to send {amount}",
-  "@scanToSend": {
-    "placeholders": {
-      "amount": {
-        "type": "String"
-      }
-    }
-  },
+  "scanToReceive": "Scan QR code to receive",
+  "@scanToReceive": {},
+  "scanToSend": "Scan QR code to pay",
+  "@scanToSend": {},
   "searchPlaceholder": "Search...",
   "@searchPlaceholder": {},
   "securitySectionTitle": "SECURITY AND PRIVACY",
@@ -428,8 +408,6 @@
   "@shareInstructions": {},
   "shareIntroFt": "Hi!\n\nI’ve just sent you ",
   "@shareIntroFt": {},
-  "shareLink": "Share payment link",
-  "@shareLink": {},
   "sharePaymentRequestLinkInstructions": "To pay me with one click, visit:",
   "@sharePaymentRequestLinkInstructions": {},
   "sharePaymentRequestLinkIntro": "Hi!\n\nI have sent a request for ",
@@ -445,10 +423,6 @@
       }
     }
   },
-  "sharePaymentRequestLinkTitle": "Share Link",
-  "@sharePaymentRequestLinkTitle": {},
-  "sharePaymentRequestQrCodeTitle": "Share QR Code",
-  "@sharePaymentRequestQrCodeTitle": {},
   "sharePaymentRequestTitle": "Request has been created!",
   "@sharePaymentRequestTitle": {},
   "shareText": "Hi!\n\nI've just sent you {amount}.\n\nTo receive it, you need to click on the link:\n\n{firstLink}",
@@ -593,8 +567,6 @@
   "@yourFirstNamePlaceholder": {},
   "yourLastNamePlaceholder": "Last Name",
   "@yourLastNamePlaceholder": {},
-  "yourLinkIsReady": "Your link is ready",
-  "@yourLinkIsReady": {},
   "yourRecoveryPhrase": "Your Secret Recovery Phrase",
   "@yourRecoveryPhrase": {},
   "yourRecoveryPhraseSub": "This phrase is composed of randomly generated 12 words. Please write down your following secret recovery phrase and save it carefully.",
