+++ packages/espressocash_app/lib/l10n/intl_fr.arb
@@ -571,22 +571,10 @@
   "@retry": {},
   "save": "Sauvegarder",
   "@save": {},
-  "scanToReceive": "Scannez pour recevoir {amount}",
-  "@scanToReceive": {
-    "placeholders": {
-      "amount": {
-        "type": "String"
-      }
-    }
-  },
-  "scanToSend": "Scanner pour envoyer {amount}",
-  "@scanToSend": {
-    "placeholders": {
-      "amount": {
-        "type": "String"
-      }
-    }
-  },
+  "scanToReceive": "Scannez pour recevoir",
+  "@scanToReceive": {},
+  "scanToSend": "Scanner pour envoyer",
+  "@scanToSend": {},
   "search": "Recherche",
   "@search": {},
   "searchPlaceholder": "Recherche...",
@@ -625,10 +613,6 @@
   "@shareInstructions": {},
   "shareIntroFt": "Hello!\n\nJe viens de t'envoyer ",
   "@shareIntroFt": {},
-  "shareLink": "Partagez le lien de paiement",
-  "@shareLink": {},
-  "shareLinks": "Partager le lien",
-  "@shareLinks": {},
   "sharePaymentRequestLinkInstructions": "Pour me payer en un seul clic, visite:",
   "@sharePaymentRequestLinkInstructions": {},
   "sharePaymentRequestLinkIntro": "Hello!\n\nJe t'envoie une demande de paiement ",
@@ -644,10 +628,6 @@
       }
     }
   },
-  "sharePaymentRequestLinkTitle": "Partager le lien",
-  "@sharePaymentRequestLinkTitle": {},
-  "sharePaymentRequestQrCodeTitle": "Partager le code QR",
-  "@sharePaymentRequestQrCodeTitle": {},
   "sharePaymentRequestTitle": "La demande a été créée!",
   "@sharePaymentRequestTitle": {},
   "shareStep1": "Étape 1 - Clique sur le premier lien:",
@@ -898,8 +878,6 @@
   "@yourEmailPlaceholder": {},
   "yourFirstNamePlaceholder": "Choisissez un nom d'utilisateur",
   "@yourFirstNamePlaceholder": {},
-  "yourLinkIsReady": "Votre lien est prêt",
-  "@yourLinkIsReady": {},
   "yourRecoveryPhrase": "Votre phrase de récupération secrète",
   "@yourRecoveryPhrase": {},
   "yourRecoveryPhraseSub": "Cette phrase est composée de 12 mots générés aléatoirement. Veuillez noter votre phrase secrète de récupération suivante et la sauvegarder attentivement.",
